+++ spec/spec_helper.rb
@@ -16,6 +16,7 @@ require 'rr'
 
 require_relative 'support/sample_fake'
 require_relative 'support/fake_creator_of_fakes'
+require_relative 'support/matchers'
 
 RSpec.configure do |config|
   config.color_enabled = true