+++ spec/support/matchers.rb
@@ -1,0 +1,3 @@
+def be_a_default_return_value
+  be_an_instance_of(Bogus::UndefinedReturnValue)
+end